requirements.txt
streamlit
pandas
requests
python-dotenv
scikit-learn